extract helper functions from the pie chart ready made
