+++ src/components/shared/Skeleton/Skeleton.module.css
@@ -18,6 +18,5 @@
   background-size: 200% 100%;
   height: 100%;
   width: 100%;
-  /* TODO: add specifics */
-  border-radius: 12px;
+  border-radius: var(--em-card-border-radius-default, 2rem);
 }
